updated dependencies

+++ update_submodules.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
-modules=( vagrant queue diff lemmatizer frontend )
+modules=(vagrant queue diff lemmatizer frontend)
 for i in "${modules[@]}"
 do
-   cd $i
-   git reset --hard
-   git pull
-   cd ../
+    if [[ $i ]]; then
+        cd $i
+        git reset --hard
+        git pull origin master
+        cd ../
+    fi
 done
